Refactor hover data management - Moved hover data structures and templates to a new module for better organization. Updated relevant imports across visualization components to utilize the new hover data module, enhancing code clarity and maintainability.
